Remove unused settings code Replace PreferenceFragment with Fragment
